backend/requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.0
celery==5.3.6
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.0
defusedxml==0.7.1
diff-match-patch==20230430
dj-database-url==2.1.0
Django==4.2.8
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-cors-headers==4.3.1
django-import-export==3.3.1
django-jazzmin==2.6.0
django-timezone-field==6.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
docopt==0.6.2
drf-yasg==1.21.7
et-xmlfile==1.1.0
gunicorn==21.2.0
idna==3.4
inflection==0.5.1
kombu==5.3.4
MarkupPy==1.14
mssql-django==1.3
mysqlclient==2.2.0
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.1
Pillow==10.0.1
pipreqs==0.4.13
prompt-toolkit==3.0.41
PyJWT==2.8.0
pyodbc==5.0.1
python-crontab==3.0.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
rest-framework-simplejwt==0.0.2
six==1.16.0
sqlparse==0.4.4
tablib==3.5.0
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
vine==5.1.0
wcwidth==0.2.12
xlrd==2.0.1
xlwt==1.3.0
yarg==0.1.9
django-cors-headers
